Move required fields to one member
